--- week3/recommend-system/src/components/Recommend/Article/Article.style.js
@@ -1,21 +1,0 @@
-import styled from "styled-components";
-
-export const ArticleContainer = styled.div`
-  display: flex;
-  justify-content: center;
-  width: 50%;
-  gap: 5%;
-  margin: 5%;
-`;
-
-export const Article = styled.div`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-`;
-
-export const StartButton = styled.button`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-`;
