+++ week3/recommend-system/src/components/Recommend/Recommend.style.js
@@ -1,13 +1,24 @@
 import styled from "styled-components";
 
-export const Container = styled.section`
+export const RecommendContainer = styled.div`
   display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: space-around;
-  width: 70%;
+  justify-content: center;
+  width: 50%;
   height: 100%;
+  gap: 5%;
+  margin: 5%;
   background-color: #ffdab9;
   margin: 5% 0;
   border-radius: 5px;
+`;
+
+export const Title = styled.div`
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  background-color: #ffefd5;
+  width: 50%;
+  padding: 0.5% 2%;
+  border-radius: 5px;
+  margin-top: 5%;
 `;
