+++ week3/recommend-system/src/components/Recommend/SelectType/index.jsx
