--- week3/recommend-system/src/components/Recommend/Title/Title.style.js
@@ -1,12 +1,0 @@
-import styled from "styled-components";
-
-export const Title = styled.div`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  background-color: #ffefd5;
-  width: 50%;
-  padding: 0.5% 2%;
-  border-radius: 5px;
-  margin-top: 5%;
-`;
