--- week3/recommend-system/src/components/Recommend/Title/index.jsx
@@ -1,6 +1,0 @@
-import * as S from "./Title.style";
-
-function Title() {
-  return <S.Title>원하는 취향을 선택해주세요</S.Title>;
-}
-export default Title;
