+++ week3/recommend-system/src/constants/Title.constant.js
@@ -1,0 +1,6 @@
+export const TitleContent = {
+  start: "원하시는 종류를 선택해주세요",
+  first: "어떤 맛을 좋아해?",
+  second: "조금 더 자세히 알려줘!",
+  third: "오리지날? 아님 신상?",
+};
